requirements.txt
Flask
pytorch
torchvision
torchaudio
Pillow
numpy
tensorflow
keras
ultralytics
Werkzeug
opencv-python
pandas